requirements.txt
PyOpenGL
PyOpenGL-accelerate
PyQt5
PyQt5-tools
PyWavefront
Numpy